requirements.txt
nltk
pandas
streamlit
spotipy
python-dotenv
scikit-learn
numpy
transformers
torch
matplotlib
